util holds some simple utility functions (big surprise)
The Scraper class is a subclass of Thread and is used to fetch all the art in the background, adding them to a Queue when ready. The items in the Queue are read by a consumer thread launched by the main application thread (reserved for the GUI). This is so the GUI can update while the artworks are retrieved without blocking and freezing