refactor to use selenium-wire over regular selenium, this greatly speeds up obtaining Sotheby's artworks by letting us inspect the GraphQL calls that Sotheby's makes, rather than rendering the whole page and traversing the DOM
also store page max value in a local file named sothpmax, binary array where each index in the array corresponds to an entry in the Category enum, and the value being the maximum number of pages for that category

also update the return type for scrape_sothebys to check whether there was enough pages to satisfy the amount of items requested. Otherwise, the amount returned must be adjusted. For example, if the user requested 7 works but only 4 are available, then that is the new limit for the game and we continue on to playing it